requirements.txt
flask
mysql-connector-python
jinja2
